--- 1/Makefile_backup
@@ -1,43 +1,0 @@
-#
-# To run the demos when linked with a shared library (default) ensure that
-# libcrypto is on the library path.  For example:
-#
-#    LD_LIBRARY_PATH=../.. ./aesccm
-
-TESTS = aesccm \
-        aesgcm \
-        aeskeywrap \
-        ariacbc	\
-		aes_encrypt \
-		aes_decrypt \
-		aria_encrypt \
-		aria_decrypt
-
-CFLAGS  = -I../../include -g -Wall
-LDFLAGS = -L../..
-LDLIBS  = -lcrypto
-
-all: $(TESTS)
-
-aesccm: aesccm.o
-aesgcm: aesgcm.o
-aeskeywrap: aeskeywrap.o
-ariacbc: ariacbc.o
-aes_encrypt: aes_encrypt.o
-aes_decrypt: aes_decrypt.o
-aria_encrypt: aria_encrypt.o
-aria_decrypt: aria_decrypt.o
-
-$(TESTS):
-	$(CC) $(CFLAGS) $(LDFLAGS) -o $@ $< $(LDLIBS)
-
-clean:
-	$(RM) $(TESTS) *.o
-
-.PHONY: test
-test: all
-	@echo "\nCipher tests:"
-	@set -e; for tst in $(TESTS); do \
-		echo "\n"$$tst; \
-		LD_LIBRARY_PATH=../.. ./$$tst; \
-	done
