+++ 1/aria_encrypt.c
@@ -21,24 +21,46 @@
 #include <openssl/err.h>
 #include <openssl/evp.h>
 
-#define KEY_SIZE 32
+#define KEY_SIZE 16
 #define IV_SIZE 12
 #define AAD_SIZE 16
 #define TAG_SIZE 16
 #define MAX_PATH_LEN 1024
 
-static uint64_t cpucycles(void) {
-    uint64_t hi, lo;
-    __asm__ __volatile__ ("CPUID\n\trdtsc\n\t" : "=a" (lo), "=d"(hi));
-    return ((uint64_t)lo) | (((uint64_t)hi) << 32);
+static struct timespec start, end;
+//nanoseconds(1e-9)
+static uint64_t wal_time_tick;
+
+static inline void INIT_WALL_TIME(void)
+{
+	wal_time_tick = 0;
+}
+
+static inline void START_WALL_TIME(void)
+{
+	clock_gettime(CLOCK_REALTIME, &start);
+}
+
+static inline void END_WALL_TIME(void)
+{
+	long seconds;
+	long nanoseconds;
+
+	clock_gettime(CLOCK_REALTIME, &end);
+
+	seconds = end.tv_sec - start.tv_sec;
+	nanoseconds = end.tv_nsec - start.tv_nsec;
+
+	wal_time_tick += ((uint64_t)seconds *(uint64_t)(1e+9) + (uint64_t)nanoseconds);
 }
 
 int main() {
-    //uint64_t start = 0, end = 0, ret = 0;
-    int counter = 1;
+    int counter = 0;
 	unsigned int calibration, tMin = 0xFFFFFFFF, start, end, entire_size = 0;
     float ret = 0;
 	calibration = calibrate();
+    INIT_WALL_TIME();
+
 
     unsigned char gcm_key[KEY_SIZE];
     unsigned char gcm_iv[IV_SIZE];
@@ -119,7 +141,9 @@ int main() {
 
 			tMin = 0xFFFFFFFF;
 			start = HiResTime();
-            //start = cpucycles();
+
+            START_WALL_TIME();
+
             /*암호화 시작*/
             EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
             if (ctx == NULL) {
@@ -181,9 +205,10 @@ int main() {
             }
 
             EVP_CIPHER_CTX_free(ctx);
-            // end = cpucycles();
-            // ret += end-start;
+
             /*암호화 끝*/
+            END_WALL_TIME();
+            
             end = HiResTime();
             if (tMin > end - start - calibration)       /* keep only the minimum time */
 			tMin = end - start - calibration;
@@ -199,13 +224,10 @@ int main() {
             fwrite(encrypted_data, 1, ct_size, fp);
             fwrite(gcm_tag, 1, TAG_SIZE, fp);
 
-            if(counter <=0){
-                printf("[%d] %7.2f cycles/byte %ld\n", KEY_SIZE, get_cpb(tMin, img_size), img_size);
-                ret += get_cpb(tMin, img_size);
-                entire_size +=img_size;
+            ret += get_cpb(tMin, img_size);
+            entire_size +=img_size;
 
-            }
-            counter--;
+            counter++;
             fclose(fp);
             free(img_data);
             free(encrypted_data);
@@ -213,9 +235,13 @@ int main() {
     }
     closedir(dir);
 
-	printf("[%d] %7.2f cycles/byte\n", KEY_SIZE, get_cpb(tMin, entire_size));
-	printf("finish. %d\n", tMin);
-	printf("finish. %.2f\n", ret/(-counter+1));
-    printf("%d\n",-counter+1);
+    double data_size = entire_size * 8;
+    double walltime = wal_time_tick / 1e9;
+    double Throughput =  data_size / walltime / 1e9;
+
+    printf("========== AES-128-GCM Encryption ==========\n");
+    printf("[*] clocks/byte: %10.6f\n", ret/counter);
+    printf("[*] walltime:    %10.6lf\n", walltime);
+    printf("[*] Throughput:  %10.6lf\n", Throughput);
     return 0;
 }
